+++ src/Domain/Segment.py
@@ -1,10 +1,11 @@
 from enforce_typing import enforce_types
+from decimal import Decimal
 
 
 @enforce_types
 class Segment:
     def __init__(self: object, id: int, name: str, 
-                 taxes: float, fees: float):
+                 taxes: Decimal, fees: Decimal):
         self.id = id
         self.name = name
         self.taxes = taxes
